+++ zedd-app/webpack.renderer.config.js
@@ -12,14 +12,14 @@ module.exports = {
     target: 'electron-renderer',
     resolve: {
         extensions: ['.js', '.ts', '.jsx', '.tsx', '.css'],
+        alias: {
+            '@aws-sdk/client-s3': false,
+        },
     },
     externals: {
         'puppeteer': 'commonjs puppeteer',
-
         'selenium-webdriver': 'commonjs selenium-webdriver',
-
         'selenium-webdriver/chrome': 'commonjs selenium-webdriver/chrome',
-
         'win-ca': 'commonjs win-ca',
     },
 }